Brought privileges table structure album_id in line with albums table change

+++ php/database/privileges_table.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS `?` (
       `user_id` int(11) NOT NULL,
-      `album_id` int(11) NOT NULL,
+      `album_id` bigint(14) unsigned NOT NULL,
       `view` tinyint(1) NOT NULL DEFAULT '0',
       `upload` tinyint(1) NOT NULL DEFAULT '0',
       `erase` tinyint(1) NOT NULL DEFAULT '0',
